fix(playback): reject absolute seek admission during ads
